% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argument-class.R
\docType{class}
\name{Argument}
\alias{Argument}
\title{Argument}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Argument
}
\description{
This class represent the arguments of an openeo backend process. It serves two purposes:
First it is used to describe the functions argument for the "describe" webservice calls, but
it is also able to store the values that are passed to the backend for execution
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the argument}

\item{\code{description}}{A short description of what the argument is about}

\item{\code{required}}{A flag whether or not this argument is required for process execution or not}

\item{\code{value}}{The actual value that is used for processing}
}}

\keyword{datasets}
