% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Band-class.R
\docType{data}
\name{Band}
\alias{Band}
\title{Band}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Band
}
\description{
This class holds the metadata information of bands for Granules and Products
}
\keyword{datasets}
