% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\docType{data}
\name{Collection}
\alias{Collection}
\title{Collection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Collection
}
\description{
This class represents the collections, which contain a set of Granules. Collections are also used to transfer
intermediate results from one process to an other (the result of a process is a Collection).
}
\section{Fields}{

\describe{
\item{\code{dimensions}}{A dimensionality object containing information about the existence of dimensions}

\item{\code{space}}{A tibble containing an index and the geometry}
}}

\keyword{datasets}
