% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection_old-class.R
\docType{data}
\name{Collection_old}
\alias{Collection_old}
\title{Collection}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Collection_old
}
\description{
This class represents the collections, which contain a set of Granules. Collections are also used to transfer
intermediate results from one process to an other (the result of a process is a Collection).
}
\section{Fields}{

\describe{
\item{\code{granules}}{A list of Granules that shall be sorted by time ascending}

\item{\code{dimensions}}{A dimensionality object containing information about the existence of dimensions}
}}

\keyword{datasets}
