% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server-class.R
\name{createServerInstance}
\alias{createServerInstance}
\title{Creates a server instance}
\usage{
createServerInstance()
}
\description{
The function creates a new server instance on the global variable 'openeo.server'. The names for
this variable is reserved and should not be changed by any means. It will crash the system, since
many endpoints will be accessing and depending on the correctly set variable 'openeo.server'.
}
