% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server-class.R
\docType{data}
\name{DatabaseEntity}
\alias{DatabaseEntity}
\title{Interface for all elements stored in a database}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DatabaseEntity
}
\description{
The class does not offer much functionality, but reserves some function names to be available to all
inheriting classes. Usually the ineriting class should implement those functions in a proper way.
}
\keyword{datasets}
