% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\docType{data}
\name{errors}
\alias{errors}
\title{Error codes and messages}
\format{a tibble}
\usage{
data(errors)
}
\description{
This data set contains the error codes and messages that shall be used in the openEO back-ends
}
\keyword{datasets}
