% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Granule-class.R
\docType{data}
\name{Granule}
\alias{Granule}
\title{Granule}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Granule
}
\description{
Smalles unit of an image collection in openEO sense. The granule is the snapshot of a part of the Earth at a given 
point in time.
}
\section{Fields}{

\describe{
\item{\code{time}}{The POSIXt object when the image was created}

\item{\code{extent}}{The spatial extent}

\item{\code{srs}}{Spatial Reference System as defined by the crs() function in the raster package}

\item{\code{data}}{Raster* object which holds the data}

\item{\code{bands}}{Named List of Band objects}
}}

\keyword{datasets}
