% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\name{importCollection}
\alias{importCollection}
\title{Imports a collection}
\usage{
importCollection(path, fun = brick)
}
\description{
The function uses a predefined lookup table for data retrieval and a metadata json document to 
describe the whole collection.
}
\details{
The files are named lookup.csv and md.json. The first has the columns "file_access", "timestamp",
"band_index" and "filename". With file_access we state the link to the accessed file how the
data will be loaded and accessed, e.g. a VRT. "filename" links to the actual file, which should be a single
band. This will be used to extract relevant metadata for each individual band (like the actual spatial 
resolution), which might not be correctly obtainable from a VRT.
}
