% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapServerConfig-class.R
\name{is.MapServerConfig}
\alias{is.MapServerConfig}
\title{Checks an object is a MapServerConfig}
\usage{
is.MapServerConfig(obj)
}
\arguments{
\item{obj}{an object to be checked}
}
\value{
logical if it is a MapServerConfig object
}
\description{
Calls the class operation on the object and checks if it is of type MapServerConfig
}
