% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapServerLayer-class.R
\name{is.MapServerLayer}
\alias{is.MapServerLayer}
\title{Checks an object is a MapServerLayer}
\usage{
is.MapServerLayer(obj)
}
\arguments{
\item{obj}{an object to be checked}
}
\value{
logical if it is a MapServerLayer object
}
\description{
Calls the class operation on the object and checks if it is of type MapServerLayer
}
