% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapServerMap-class.R
\name{is.MapServerMap}
\alias{is.MapServerMap}
\title{Checks an object is a MapServerMap}
\usage{
is.MapServerMap(obj)
}
\arguments{
\item{obj}{an object to be checked}
}
\value{
logical if it is a MapServerMap object
}
\description{
Calls the class operation on the object and checks if it is of type MapServerMap
}
