% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapServerWeb-class.R
\name{is.MapServerWeb}
\alias{is.MapServerWeb}
\title{Checks an object is a MapServerWeb}
\usage{
is.MapServerWeb(obj)
}
\arguments{
\item{obj}{an object to be checked}
}
\value{
logical if it is a MapServerWeb object
}
\description{
Calls the class operation on the object and checks if it is of type MapServerWeb
}
