% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server-class.R
\name{migrateFilesToDB}
\alias{migrateFilesToDB}
\title{Migration tool}
\usage{
migrateFilesToDB(db, workspace)
}
\arguments{
\item{db}{sqlite database connection}

\item{workspace}{the user workspace, where to find the user folders}
}
\description{
This functions migrates the old file based information storage into the sqlite solution.
}
