% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process-class.R
\docType{class}
\name{Process}
\alias{Process}
\title{Process}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Process
}
\description{
The process class in the openEO thinking. It is a described function that can be serialized in short and detail for the webservice, 
but also relates to the internal R function (process) that will be applied to perform the task.
}
\details{
The class is used to describe a processing entity, which means it stores its description (textual), the described
Arguments (see Argument) and the R operation. The class is mainly used as a template class, but it can be made executable. 
When the process is executable then the arguments will have values assigned to them.
}
\section{Fields}{

\describe{
\item{\code{process_id}}{The id or name of the process which will be used to be addressed from the webservice}

\item{\code{args}}{The list of arguments for the underlying function}

\item{\code{description}}{The textual description of the process}

\item{\code{dimensions_modifier}}{A DimensionModifier class that will be applied to the input data to describe how the data is changed during
processing.}

\item{\code{operation}}{A function that executes the operation in R described by this process}
}}

\keyword{datasets}
