% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Generic-class.R
\docType{data}
\name{R2Generic}
\alias{R2Generic}
\title{R2Generic}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
R2Generic
}
\description{
This class prepares objects of class `Collection` to be written to disk in a generic format (GeoTIFF) for 
the User-Defined Functions(UDFs) to run. Apart from the binary raster files, it also writes an ASCII "legend" 
in CSV format which acts as a look-up table containing metadata on the individual rasters. This look-up table 
is parsed to load the rasters into a `stars` object from the function `run_UDF()` in the package `openEO.R.UDF`.
}
\section{Fields}{

\describe{
\item{\code{granules}}{Object of class Granule}
}}

\keyword{datasets}
