% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UdfRuntime-class.R
\name{r_filebased_udf_runtime}
\alias{r_filebased_udf_runtime}
\title{UDF runtime filebased}
\usage{
r_filebased_udf_runtime()
}
\description{
Creates a UDF runtime that is able to run the UDF in the R local filebased approach.
}
