% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UdfRuntime-class.R
\name{r_udf_service_runtime}
\alias{r_udf_service_runtime}
\title{Create the service runtime}
\usage{
r_udf_service_runtime()
}
\description{
Creates a UDFRuntime object for the server in order to run an UDF on the R-UDF webservice.
}
