% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_UDF.R
\name{raster_collection_export}
\alias{raster_collection_export}
\title{Creates RasterCollectionTile representation}
\usage{
raster_collection_export(collection)
}
\arguments{
\item{collection}{Collection object}
}
\value{
list that can be transformed into "UdfData" JSON
}
\description{
Subsets and groups Collection data by band and space in order to create the specified UDF RasterCollectionTile JSON output.
}
