% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality.R
\name{read_dimensionality}
\alias{read_dimensionality}
\title{Reads dimensionality from a binary code string}
\usage{
read_dimensionality(code)
}
\arguments{
\item{code}{code string or integer value that represents a value in 0 to 2^5-1 at most}
}
\value{
Dimensionality object
}
\description{
The function reads and interpretes a code string containing 1 and 0 for each dimension.
}
