% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R
\name{read_legend}
\alias{read_legend}
\title{Reads a legend file and creates a Collection}
\usage{
read_legend(legend.path, code, ...)
}
\arguments{
\item{legend.path}{Absolute file path to the legend file}

\item{code}{A dimensionality code or an integer that represents the binary code}

\item{...}{contains attributes that are passed on to addGranule as meta_band}
}
\value{
a Collection object
}
\description{
The function uses a legend.csv file as described in 
https://github.com/pramitghosh/OpenEO.R.UDF/blob/master/data/example_udf_out/out_legend.csv and a dimensionality code
as well as optional additional parameters for band metadata.
}
