% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializer_proxy.R
\name{serializer_proxy}
\alias{serializer_proxy}
\title{Proxy serializer}
\usage{
serializer_proxy()
}
\value{
a serializer function that passes through the response - as is - from the redirected server
}
\description{
This serializer will serialize the data of a httr response. The body will become content(val) and the
content-type is set accordingly. Also the status code will be transferred from httr-response to the httpuv response.
}
