% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServerConfig-class.R
\name{ServerConfig}
\alias{ServerConfig}
\title{Server Configuration}
\usage{
ServerConfig()
}
\description{
The server configurations
}
\section{Fields}{

\describe{
\item{\code{baseserver.url}}{The URL of where this server is accesible. Trailing slashes are removed.}
}}

