% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection-class.R, R/prepare_UDF.R
\name{udf_export}
\alias{udf_export}
\alias{udf_export}
\title{Function to export a collection for the UDF webservice}
\usage{
udf_export(collection)

udf_export(collection)
}
\arguments{
\item{collection}{Collection object}
}
\value{
list that can be transformed into "UdfData" JSON
}
\description{
The function currently transforms the data and extent into lists that can be transformed with toJSON() into
the required JSON object for "data"

Transforms the data contained in a Collection into a JSON representation. It will be passed along the code script URL as data
to the specified UDF REST processing service. Currently implemented only for raster timeserires collections.
}
