% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UdfRuntime-class.R
\name{UdfRuntime}
\alias{UdfRuntime}
\title{A constructor for UdfRuntime class}
\usage{
UdfRuntime(id, description = NULL, language, version = NULL,
  libraries = NULL, docker = NULL, tag = NULL, links = NULL)
}
\arguments{
\item{id}{the id at which the udf runtime will be referencable at openeo.server$udf_runtimes}

\item{description}{a textual description about the udf runtime environment}

\item{language}{the underlying programming language to execute the udf script in}

\item{version}{the version of the programming langunge}

\item{libraries}{a list of installed libraries of the programming language}

\item{docker}{the docker image name (either on dockerhub or on a private repository)}

\item{tag}{the tag of the docker image}

\item{links}{a link to point to additional information}
}
\value{
UdfRuntime class
}
\description{
A S3 class to store all relevant information about a udf runtime environment for visualization purposes to a user (used 
at /udf_runtimes endpoint)
}
