% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Generic-class.R
\name{write_generics}
\alias{write_generics}
\title{Write generics to disk}
\usage{
write_generics(collection_obj, dir_name = "disk", logger)
}
\arguments{
\item{collection_obj}{Object of class Collection as produced in the previous step while executing the process graph before encountering the UDF}

\item{dir_name}{Path where the generics are to be written to disk. This could be obtained from the UDF process if it is defined by the user while registering it.}
}
\description{
This function writes rasters in a consistent directory structure to disk in generic GeoTIFF formats. 
It takes as input an object of class `Collection` and the path where the files are to be written to disk. Once the files 
have been written to disk, it can be loaded into a `stars` object by the user after which custom functions could 
be applied to it.
}
