% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{data}
\name{Job}
\alias{Job}
\title{Job}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Job
}
\description{
This class represents an openEO job. Which is submitted by an user, containing an executable process graph. It is strongly bound
to the backend, since the OpenEOServer class needs to be accessible, when it comes to loading related process for the process 
graph.
}
\section{Fields}{

\describe{
\item{\code{job_id}}{The unique identifier of the job}

\item{\code{status}}{The current status in the job lifecycle}

\item{\code{process_graph}}{graph of nested processes that is executable (ExecutableProcess)}

\item{\code{view}}{Spatio-Temporal Extent to be used for the calculation (currently not in use)}

\item{\code{submitted}}{Timestamp when the job was submitted to the server}

\item{\code{user_id}}{The user who owns the job}

\item{\code{consumed_credits}}{For accounting and billing the amount of credits consumed by this job}

\item{\code{last_update}}{Timestamp when the job was last updated}

\item{\code{output}}{output configuration like output$format copied from the process graph (pg)}

\item{\code{results}}{Contains the result of the process_graph after execution (Collection)}

\item{\code{persistent}}{Whether or not the job is stored in database}

\item{\code{output}}{list containing the output configuration like format (or additional GDAL commands)}
}}

\keyword{datasets}
