% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_UDF.R
\name{udf_export}
\alias{udf_export}
\title{Prepares the collection data for the UDF service request}
\usage{
udf_export(collection, strategy)
}
\arguments{
\item{collection}{Collection object}

\item{strategy}{the tiling strategy (not implemented yet)}
}
\value{
list that can be transformed into "UdfData" JSON
}
\description{
Transforms the data contained in a Collection into a JSON representation. It will be passed along the code script URL as data
to the specified UDF REST processing service. Currently implemented only for raster timeserires collections.
}
